Refactor color palette management into shared module
